+++ gen/DocumentFormat.OpenXml.Generator/PackageInformation.cs
@@ -1,0 +1,10 @@
+﻿// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+using DocumentFormat.OpenXml.Generator.Models;
+
+namespace DocumentFormat.OpenXml.Generator;
+
+public record PackageInformation(string ClassName, IEnumerable<Part> Parts)
+{
+}
